# openapi.yaml
openapi: 3.0.3
info:
  title: API ИИ-гида музея Фаберже
  version: 0.1.0-mvp
  description: |
    Backend-API мобильного web-приложения (PWA) «ИИ-гид музея Фаберже».

    Посетитель сканирует **QR-код** на входе экспозиции, попадает в приложение
    без регистрации и установки, перемещается по **интерактивной карте**
    (зал → витрина → экспонат), а главное — может **сфотографировать экспонат**,
    получить о нём рассказ от ИИ-гида и **прослушать** его озвучку.

    Этот документ описывает контракт backend-сервиса (FastAPI на Yandex Cloud),
    к которому обращается фронтенд. Спецификация написана по дорожной карте MVP
    (25 мая – 25 июня) и служит единым контрактом между Frontend, Backend и ML.

    ## Ключевой пользовательский сценарий (бесшовное распознавание)

    ```
    Клик «Распознать экспонат»
      → камера (WebRTC)
      → снимок
      → POST /recognition (фото → YOLO → label_slug)
      → POST /guide/story (label_slug → raw_history из БД → YandexGPT → рассказ)
      → показ рассказа + 3–4 всплывающих вопроса
      → POST /speech («Прослушать», SpeechKit → аудио)
      → блок «Другие экспонаты зала» (GET /exhibits/{id}/related)
    ```

    ## Доменная модель

    Иерархия каталога повторяет физическую структуру музея:

    | Уровень   | Сущность   | Таблица БД   | Ключевые поля                                   |
    |-----------|------------|--------------|-------------------------------------------------|
...

...
    (через публичный API не отдаётся).

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security: []
paths:
...
  /halls:
    get:
      tags: [Карта и навигация]
      summary: Список залов
      description: >-
        Постраничный список залов музея для карты и навигации. Служебные записи
        (`is_service`) в выдачу не попадают — источник правды один, скрывать их
        на клиенте не нужно. Описание зала приходит и полным (`description`), и
        превью (`description_preview` + `description_has_more`) — см.
        `GET /halls/{hall_id}`.
      operationId: listHalls
      security: []
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/IsTemporary'
        - $ref: '#/components/parameters/IncludeService'
      responses:
        '200':
          description: Список залов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HallListResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /recognition:
    post:
      tags: [Распознавание]
      summary: Распознать экспонат по фото
      description: |
        Принимает фотографию экспоната, отправляет её в развёрнутую на
        Yandex Cloud модель **YOLO** и возвращает распознанный `label_slug`
        вместе с найденной в БД карточкой экспоната.

        Сразу после снимка фронтенд бесшовно переходит в окно чата с
        состоянием загрузки и вызывает этот эндпоинт. При успешном
        распознавании далее обычно вызывается `POST /guide/story`.

        Если уверенность ниже порога — `recognized = false`; в `candidates`
        могут вернуться наиболее вероятные классы для уточнения.

        **Лимит на файл — 2,5 МБ** (`MAX_UPLOAD_MB`), превышение даёт 413 с
        текстом «Фото слишком большое — X МБ при допустимых 2,5 МБ». Он
        намеренно ниже платформенной границы Yandex API Gateway — **3,5 МиБ
        (3 670 016 байт) на весь запрос**, столько же на payload синхронного
        вызова Cloud Functions, куда тело уезжает в base64 (+33 %). Это не наш
        выбор и кодом не поднимается — только конфигурацией API Gateway.

        Почему граница важна для фронта: гейтвей режет запрос ДО приложения и
        отвечает своим 413 без CORS-заголовков, поэтому в браузере это не «фото
        слишком большое», а «Failed to fetch». Пока наш лимит ниже
        платформенного, 413 отдаёт сам сервис — с CORS и внятным текстом
        (п. 10 баг-репорта 06.08.2026). Кадр с камеры телефона (3–6 МБ) нужно
        сжимать перед отправкой; кадр из canvas на экране распознавания
        (~0,5 МБ) проходит как есть.
      operationId: recognizeExhibit
      security: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Фото экспоната (JPEG/PNG/WebP, до 2,5 МБ).
                hall_id:
                  type: integer
                  minimum: 1
                  description: >-
                    Необязательная подсказка о текущем зале (по карте) —
                    повышает точность и сужает поиск.
                top_k:
                  type: integer
                  minimum: 1
                  maximum: 10
                  default: 3
                  description: Сколько кандидатов вернуть при неуверенном распознавании.
            encoding:
              file:
                contentType: image/jpeg, image/png, image/webp
      responses:
        '200':
          description: Результат распознавания (успех или «не распознано»).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecognitionResponse'
              examples:
                recognized:
                  summary: Экспонат распознан
                  value:
                    recognized: true
                    label_slug: faberge_egg_coronation
                    confidence: 0.94
                    request_id: 7b1c0e2a-9f3d-4a2b-8c10-2a4d6e8f0a11
                    processing_ms: 312
                    exhibit:
                      id: 101
                      label_slug: faberge_egg_coronation
                      name: Яйцо «Коронационное»
                      image_url: https://cdn.example.cloud/exhibits/coronation/main.jpg
                      location:
                        hall_id: 4
                        hall_number: 4
                        hall_name: Синяя гостиная
                        showcase_id: 12
                        showcase_number: 5
                        showcase_name: Пасхальные подарки
                        text: Зал 4 «Синяя гостиная», витрина 5
                        text_in: в зале 4 «Синяя гостиная», витрина 5
                      hall:
                        id: 4
                        hall_number: 4
                        name: Синяя гостиная
                      showcase:
                        id: 12
                        showcase_number: 5
                        name: Пасхальные подарки
                      year_created: '1897'
                      origin_place: Санкт-Петербург
                      maker:
                        text: Фирма К. Фаберже, мастер М. Перхин
                        firm: Фирма К. Фаберже
                        master: мастер М. Перхин
                      master_name: Фирма К. Фаберже, мастер М. Перхин
                      material: Золото, гильоше-эмаль, бриллианты, рубины
                      techniques: гравировка, канфарение, эмаль по гильошированному фону
                      short_description: >-
                        Пасхальное яйцо-сюрприз, подаренное Николаем II
                        императрице Александре Фёдоровне в память о коронации.
                      model_3d_url: https://koinovo.ru/fabergemuseum/coronation
                    candidates: []
                not_recognized:
                  summary: Не распознано — есть кандидаты
                  value:
                    recognized: false
                    label_slug: null
                    confidence: 0.41
                    request_id: 0c5b2a18-3e7f-4d9a-9b21-7c3e1f5a2b00
                    processing_ms: 287
                    exhibit: null
                    candidates:
                      - label_slug: faberge_egg_winter
                        name: Яйцо «Зимнее»
                        confidence: 0.41
                      - label_slug: faberge_egg_lilies
                        name: Яйцо «Ландыши»
                        confidence: 0.33
        '400':
          $ref: '#/components/responses/BadRequest'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '502':
          $ref: '#/components/responses/UpstreamError'

  # ───────────────────────────── ИИ-гид ──────────────────────────────────
...
  /guide/chat:
    post:
      tags: [ИИ-гид]
      summary: Диалог с ИИ-гидом
      description: |
        Многоходовой диалог с гидом. Контекст задаётся экспонатом
        (`exhibit_id`/`label_slug`) либо залом (`hall_id` — для кнопки
        «Спросить AI-гида» на уровне зала). Часто `message` — это один из
        ранее предложенных `suggested_questions`.

        Состояние диалога хранится по `session_id`: если он не передан, сервер
        создаёт новый и возвращает его. Альтернативно можно передавать
        историю сообщений в `history`. Ответ снова содержит набор
        контекстных вопросов-подсказок.
      operationId: chatWithGuide
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
            examples:
              followUp:
                summary: Уточняющий вопрос по экспонату
                value:
                  session_id: 2f6a9b3c-1d4e-4f8a-9c2b-5e7a1d0c3b66
                  context:
                    exhibit_id: 101
                  message: Кому подарили это яйцо?
      responses:
        '200':
          description: Ответ гида и новые подсказки.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
              examples:
                answer:
                  summary: Ответ на уточняющий вопрос
                  value:
                    session_id: 2f6a9b3c-1d4e-4f8a-9c2b-5e7a1d0c3b66
                    answer: >-
                      Это яйцо Николай II подарил своей супруге, императрице
                      Александре Фёдоровне, на Пасху 1897 года…
                    suggested_questions:
                      - А что подарили его матери?
                      - Где оно хранилось после революции?
                      - Сколько таких яиц всего создал Фаберже?
                    context:
                      exhibit_id: 101
                      label_slug: faberge_egg_coronation
                    referenced_exhibits:
                      - id: 101
                        name: Яйцо «Коронационное»
                        exhibit_number: '12'
                        thumbnail_url: https://cdn.example.cloud/exhibits/coronation/thumb.jpg
                        hall_number: 4
                        showcase_number: 5
                        mentioned_in: context
                    referenced_halls: []
                    location: null
                noMentions:
                  summary: Ответ ни на один предмет каталога не ссылается — блок «Упомянуто в ответе» не показываем
                  value:
                    session_id: 2f6a9b3c-1d4e-4f8a-9c2b-5e7a1d0c3b66
                    answer: >-
                      Музей Фаберже открылся в Шуваловском дворце в 2013 году.
                      Основу собрания составила коллекция Виктора Вексельберга…
                    # Общий чат без контекста экспоната: подсказок здесь раньше
                    # не было вовсе (п. II-7). Теперь отдаётся детерминированный
                    # набор про музей — без обращения к модели.
                    suggested_questions:
                      - Какие залы есть в музее?
                      - Что обязательно посмотреть в коллекции?
                      - Чем знаменит Карл Фаберже?
                    context: null
                    referenced_exhibits: []
                    referenced_halls: []
                    location: null
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '502':
          $ref: '#/components/responses/UpstreamError'

  # ──────────────────────────── Озвучивание ──────────────────────────────
  /speech:
    post:
      tags: [Озвучивание]
      summary: Синтез речи (Прослушать)
      description: |
        Озвучивает текст через **Yandex SpeechKit** для кнопки «Прослушать».
        Принимает либо произвольный `text`, либо `exhibit_id` (тогда озвучивается
        краткое описание экспоната). Возвращает ссылку на аудио в Object Storage
        (раздаётся через CDN); часто используемые описания кэшируются.

        Кнопка «Прослушать» на фронте имеет состояния **Play / Pause / Loading** —
        `Loading` соответствует ожиданию ответа этого эндпоинта.
      operationId: synthesizeSpeech
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpeechRequest'
            examples:
              byText:
                summary: Озвучить произвольный текст
                value:
                  text: В 1897 году Карл Фаберже исполнил Коронационное яйцо…
                  voice: alena
                  format: mp3
                  speed: 1.0
              byExhibit:
                summary: Озвучить описание экспоната
                value:
                  exhibit_id: 101
                  voice: filipp
                  format: oggopus
      responses:
        '200':
          description: Ссылка на синтезированное аудио.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpeechResponse'
              examples:
                ok:
                  summary: Готовое аудио
                  value:
                    audio_url: https://cdn.example.cloud/tts/coronation_alena.mp3
                    format: mp3
                    voice: alena
                    duration_ms: 21400
                    characters: 248
                    cached: true
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '502':
          $ref: '#/components/responses/UpstreamError'

  # ─────────────────────── Администрирование [вне MVP] ───────────────────
...
  /admin/halls:
    post:
      tags: [Администрирование]
      summary: '[Вне MVP] Создать зал'
      operationId: createHall
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HallCreate'
      responses:
        '201':
          description: Зал создан.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HallDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /admin/showcases:
    post:
      tags: [Администрирование]
      summary: '[Вне MVP] Создать витрину'
      operationId: createShowcase
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShowcaseCreate'
      responses:
        '201':
          description: Витрина создана.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShowcaseDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /admin/exhibits:
    post:
      tags: [Администрирование]
      summary: '[Вне MVP] Создать экспонат'
      description: >-
        Создание экспоната, включая внутреннее поле `raw_history`
        (факты для генерации рассказа LLM).
      operationId: createExhibit
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExhibitCreate'
            examples:
              winterEgg:
                summary: Зимнее яйцо с raw_history
                value:
                  showcase_id: 12
                  label_slug: faberge_egg_winter
                  name: Яйцо «Зимнее»
                  year_created: '1913'
                  master_name: Альма Пиль
                  material: Горный хрусталь, платина, бриллианты
                  techniques: резьба по камню, гравировка, монтировка
                  short_description: >-
                    Пасхальное яйцо в виде ледяного кристалла с сюрпризом-корзиной
                    подснежников.
                  raw_history: >-
                    Заказчик — Николай II. Подарок императрице Марии Фёдоровне.
                    Дизайн — Альма Пиль. Использованы горный хрусталь и бриллианты,
                    имитирующие иней. Сюрприз — корзина подснежников.
      responses:
        '201':
          description: Экспонат создан.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExhibitAdmin'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /admin/exhibits/{exhibit_id}:
    get:
      tags: [Администрирование]
      summary: '[Вне MVP] Карточка экспоната для админки'
      description: >-
        Полная карточка экспоната, включая внутреннее поле `raw_history`
        (факты для LLM), которое не отдаётся публичным `GET /exhibits/{id}`.
        Нужна админке для просмотра/редактирования полного описания.
      operationId: getExhibitAdmin
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExhibitId'
      responses:
        '200':
          description: Карточка экспоната (с `raw_history`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExhibitAdmin'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Администрирование]
      summary: '[Вне MVP] Обновить экспонат'
      description: >-
        Обновление карточки. Тело сливается с текущей карточкой: поле, которого
        НЕТ в теле запроса, остаётся как было, а явный `null` поле стирает —
        чтобы очистить описание или ссылку, пришлите `null`, а не пропустите поле.


        До 31.08.2026 было наоборот: непереданное поле обнулялось. Админка
        присылает форму частично, и правка техник у яйца «Ренессанс» унесла
        изображение, описание и материалы (docs/task-2026-08-31-admin-put-data-loss.md).
        Прежняя семантика полной замены доступна по `?full_replace=true`.


        Отдельное правило для `image_url`: он — зеркало строки галереи с
        `is_primary`, поэтому если поля в теле НЕТ, а в карточке оно пусто при
        целой галерее, сервер вернёт URL оттуда (лечение карточек, испорченных
        до 31.08.2026). Явно присланный `image_url` — в том числе `null` — это
        лечение отключает и записывается как есть: очистить главное фото можно.
        Правило действует и при `?full_replace=true`: `image_url` — единственное
        поле, которое полная замена не обнуляет.


        Если нужно поменять одно-два поля, правильнее `PATCH /admin/exhibits/{id}`.
      operationId: updateExhibit
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExhibitId'
        - name: full_replace
          in: query
          required: false
          description: >-
            Полная замена: поля, которых нет в теле, обнуляются. По умолчанию
            false — непереданные поля сохраняются. Исключение одно — `image_url`:
            он зеркало строки галереи с `is_primary` и восстанавливается из неё
            даже при полной замене; чтобы снять главное фото, пришлите
            `image_url: null` явно.
          schema:
            type: boolean
            default: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExhibitUpdate'
      responses:
        '200':
          description: Экспонат обновлён.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExhibitAdmin'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    patch:
      tags: [Администрирование]
      summary: '[Вне MVP] Частично обновить экспонат'
      description: >-
        Частичное обновление: пишутся только переданные поля. Явный `null`
        стирает поле — других способов очистить его нет, включая `image_url`
        (`{"image_url": null}` снимает главное фото, оставляя снимок в галерее).
        Если же `image_url` в теле НЕ передан, а в карточке он пуст при целой
        галерее, сервер подставит URL строки с `is_primary` — то же лечение, что
        и у `PUT`. Рекомендуемый способ правки карточки.
      operationId: patchExhibit
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExhibitId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExhibitPatch'
      responses:
        '200':
          description: Экспонат обновлён.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExhibitAdmin'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    delete:
      tags: [Администрирование]
      summary: '[Вне MVP] Удалить экспонат'
      operationId: deleteExhibit
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExhibitId'
      responses:
        '204':
          description: Экспонат удалён.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Статический Bearer-токен администратора (получается через
        `POST /admin/login`). Слать в заголовке `Authorization: Bearer <token>`
        ко всем эндпоинтам `/admin/**` (кроме `/admin/login`).
  parameters:
    Limit:
      name: limit
      in: query
      description: Максимум элементов на странице.
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      description: Смещение от начала выборки (постраничность).
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    IsTemporary:
      name: is_temporary
      in: query
      required: false
      description: >-
        Фильтр по типу экспозиции: `true` — временные (сменные) выставки,
        `false` — основная (постоянная) экспозиция. Без параметра возвращаются
        и те и другие. Для экспонатов признак берётся от зала (через витрину).
      schema:
        type: boolean
      example: true
    IncludeService:
      name: include_service
      in: query
      required: false
      description: >-
        Включить служебные записи каталога (`is_service`) — технические залы и
        вспомогательные пространства. По умолчанию `false` — посетителю такие
        записи не показываются. `true` нужен админке, чтобы служебный зал
        оставался управляемым.
      schema:
        type: boolean
        default: false
      example: false
...